[tint] Signed integer polyfill for IR
Bug: 426999765
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/249135
Commit-Queue: Peter McNeeley <petermcneeley@google.com>

+++ test/tint/expressions/binary/left-shift/scalar-scalar/i32.wgsl.expected.glsl
@@ -4,5 +4,5 @@ layout(local_size_x = 1, local_size_y = 1, local_size_z = 1) in;
 void main() {
   int a = 1;
   uint b = 2u;
-  int r = (a << (b & 31u));
+  int r = int((uint(a) << (b & 31u)));
 }
