+++ test/tint/statements/compound_assign/scalar/shift_left.wgsl.expected.glsl
@@ -10,7 +10,7 @@ buffer v_block_1_ssbo {
   S inner;
 } v_1;
 void foo() {
-  v_1.inner.a = (v_1.inner.a << (2u & 31u));
+  v_1.inner.a = int((uint(v_1.inner.a) << (2u & 31u)));
 }
 layout(local_size_x = 1, local_size_y = 1, local_size_z = 1) in;
 void main() {
